pyproject.toml
[project]
name = "linkedinscraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "elasticsearch>=8.11.0,<9.0.0",
    "pandas>=2.3.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "langdetect>=1.0.9",
    "openai>=1.86.0",
    "pdfminer.six>=20240706",
    "pypdf2>=3.0.0",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
]
debug = [
    "gittyleaks>=0.0.31",
    "pre-commit>=4.2.0",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
]


[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-ra -q"
